# Conflicts:
#	src/core/string_consumer.cpp
#	src/fileio_type.h
#	src/fios.cpp
#	src/fios.h
#	src/misc/endian_buffer.hpp
#	src/network/core/packet.cpp
#	src/network/core/packet.h
#	src/network/network_admin.h
#	src/network/network_func.h
#	src/newgrf_airport.cpp
#	src/newgrf_badge.cpp
#	src/newgrf_canal.cpp
#	src/newgrf_cargo.cpp
#	src/newgrf_railtype.cpp
#	src/newgrf_roadtype.cpp
#	src/newgrf_station.cpp
#	src/openttd.cpp
#	src/pathfinder/yapf/yapf_rail.cpp
#	src/saveload/saveload.cpp
#	src/settings_gui.cpp
#	src/settings_internal.h
#	src/settingsgen/settingsgen.cpp
#	src/string_func.h
#	src/string_type.h
#	src/strings.cpp
#	src/strings_internal.h
#	src/strings_type.h
#	src/tests/string_func.cpp
#	src/video/opengl.cpp

+++ src/network/network.cpp
@@ -1288,7 +1288,7 @@ void NetworkGameLoop()
 					extern void NetworkSendCommandImplementation(Commands cmd, TileIndex tile, const CommandPayloadBase &payload, StringID error_msg, CommandCallback callback, CallbackParameter callback_param, CompanyID company);
 					NetworkSendCommandImplementation(cp->command_container.cmd, cp->command_container.tile, *cp->command_container.payload, (StringID)0, CommandCallback::None, 0, cp->company);
 					Debug(net, 0, "injecting: {}; {:02x}; {:06x}; {:08x} ({})",
-							debug_date_dumper().HexDate(), (int)_current_company, cp->command_container.tile, cp->command_container.cmd, GetCommandName(cp->command_container.cmd));
+							debug_date_dumper().HexDate(), _current_company, cp->command_container.tile, cp->command_container.cmd, GetCommandName(cp->command_container.cmd));
 					cp.reset();
 				}
 				if (check_sync_state) {
